# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CachePurgeResult": {
        "properties": {
          "purged": {
            "title": "Purged",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "purged"
        ],
        "title": "CachePurgeResult",
        "type": "object"
      },
      "CacheResyncRequest": {
        "properties": {
          "pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Single pattern to refresh; omit to re-check all cached patterns",
            "title": "Pattern"
          },
          "upstream": {
            "description": "Upstream name to resync against",
            "title": "Upstream",
            "type": "string"
          }
        },
        "required": [
          "upstream"
        ],
        "title": "CacheResyncRequest",
        "type": "object"
      },
      "CacheResyncResponse": {
        "properties": {
          "checked": {
            "title": "Checked",
            "type": "integer"
          },
          "cleared": {
            "title": "Cleared",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CacheResyncResult"
            },
            "title": "Results",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          },
          "upstream": {
            "title": "Upstream",
            "type": "string"
          }
        },
        "required": [
          "success",
          "upstream",
          "checked",
          "updated",
          "cleared",
          "results"
        ],
        "title": "CacheResyncResponse",
        "type": "object"
      },
...
      "CheckLogEntry": {
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "check_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Url"
          },
          "count": {
            "default": 1,
            "title": "Count",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "tried_at": {
            "title": "Tried At",
            "type": "string"
          },
          "upstream_name": {
            "title": "Upstream Name",
            "type": "string"
          }
        },
        "required": [
          "pattern",
          "upstream_name",
          "tried_at"
        ],
        "title": "CheckLogEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpstreamCacheEntry": {
        "properties": {
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "resolved_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Url"
          },
          "upstream_name": {
            "title": "Upstream Name",
            "type": "string"
          }
        },
        "required": [
          "pattern",
          "upstream_name",
          "checked_at"
        ],
        "title": "UpstreamCacheEntry",
        "type": "object"
      },
      "UpstreamCreate": {
        "properties": {
          "base_url": {
            "description": "Target base URL, e.g. https://go.dev",
            "title": "Base Url",
            "type": "string"
          },
          "fail_status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Status code indicating missing shortcut",
            "title": "Fail Status Code"
          },
          "fail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URL returned by upstream on 404",
            "title": "Fail Url"
          },
          "name": {
            "description": "Unique label for the upstream",
            "title": "Name",
            "type": "string"
          },
          "skip_sso_cache": {
            "default": false,
            "description": "Never cache SSO target matches",
            "title": "Skip Sso Cache",
            "type": "boolean"
          },
          "verify_ssl": {
            "default": true,
            "description": "Verify SSL certs",
            "title": "Verify Ssl",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "base_url"
        ],
        "title": "UpstreamCreate",
        "type": "object"
      },
      "UpstreamRead": {
        "properties": {
          "base_url": {
            "description": "Target base URL, e.g. https://go.dev",
            "title": "Base Url",
            "type": "string"
          },
          "fail_status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Status code indicating missing shortcut",
            "title": "Fail Status Code"
          },
          "fail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URL returned by upstream on 404",
            "title": "Fail Url"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "description": "Unique label for the upstream",
            "title": "Name",
            "type": "string"
          },
          "skip_sso_cache": {
            "default": false,
            "description": "Never cache SSO target matches",
            "title": "Skip Sso Cache",
            "type": "boolean"
          },
          "verify_ssl": {
            "default": true,
            "description": "Verify SSL certs",
            "title": "Verify Ssl",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "base_url"
        ],
        "title": "UpstreamRead",
        "type": "object"
      },
...
      "UpstreamUpdate": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "fail_status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fail Status Code"
          },
          "fail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fail Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "skip_sso_cache": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Sso Cache"
          },
          "verify_ssl": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Ssl"
          }
        },
        "title": "UpstreamUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Async FastAPI backend for Redirector (EPIC-01). Shortcuts, upstreams, resolve hot path, auth, metrics, QR.",
    "title": "Redirector API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/upstreams": {
      "get": {
        "operationId": "list_upstreams_api_v1_upstreams_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UpstreamRead"
                  },
                  "title": "Response List Upstreams Api V1 Upstreams Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List all upstreams",
        "tags": [
          "upstreams"
        ]
      },
      "post": {
        "operationId": "create_upstream_api_v1_upstreams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpstreamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpstreamRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create an upstream",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/cache": {
      "delete": {
        "operationId": "purge_upstream_cache_api_v1_upstreams_cache_delete",
        "parameters": [
          {
            "description": "Purge specific upstream only, or all if omitted",
            "in": "query",
            "name": "upstream",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Purge specific upstream only, or all if omitted",
              "title": "Upstream"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CachePurgeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Purge upstream shortcut cache",
        "tags": [
          "upstreams"
        ]
      },
      "get": {
        "operationId": "list_upstream_cache_api_v1_upstreams_cache_get",
        "parameters": [
          {
            "description": "Filter by upstream name",
            "in": "query",
            "name": "upstream",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by upstream name",
              "title": "Upstream"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UpstreamCacheEntry"
                  },
                  "title": "Response List Upstream Cache Api V1 Upstreams Cache Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List cached upstream shortcuts",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/cache/resync": {
      "post": {
        "description": "Re-check one pattern (or all cached ones) and reconcile cache rows.\n\nSynchronous by design at this scale; graduate to arq workers (EPIC-04/06)\nif resync-all outgrows the 3 s-per-check request budget.",
        "operationId": "resync_upstream_cache_api_v1_upstreams_cache_resync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheResyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheResyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resync upstream cache",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/cache/{upstream}/{pattern}": {
      "delete": {
        "operationId": "purge_cache_entry_api_v1_upstreams_cache__upstream___pattern__delete",
        "parameters": [
          {
            "in": "path",
            "name": "upstream",
            "required": true,
            "schema": {
              "title": "Upstream",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pattern",
            "required": true,
            "schema": {
              "title": "Pattern",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CachePurgeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Purge one cached upstream entry",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/check-logs": {
      "delete": {
        "operationId": "clear_check_logs_api_v1_upstreams_check_logs_delete",
        "parameters": [
          {
            "description": "Clear one upstream only, or all if omitted",
            "in": "query",
            "name": "upstream",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Clear one upstream only, or all if omitted",
              "title": "Upstream"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CachePurgeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear upstream check logs",
        "tags": [
          "upstreams"
        ]
      },
      "get": {
        "operationId": "list_check_logs_api_v1_upstreams_check_logs_get",
        "parameters": [
          {
            "description": "Filter by upstream name",
            "in": "query",
            "name": "upstream",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by upstream name",
              "title": "Upstream"
            }
          },
          {
            "description": "Max entries, newest first",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Max entries, newest first",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckLogEntry"
                  },
                  "title": "Response List Check Logs Api V1 Upstreams Check Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List upstream check logs",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/check/stream/{pattern}": {
      "get": {
        "operationId": "stream_upstream_check_api_v1_upstreams_check_stream__pattern__get",
        "parameters": [
          {
            "in": "path",
            "name": "pattern",
            "required": true,
            "schema": {
              "title": "Pattern",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream real-time upstream resolution check via SSE",
        "tags": [
          "upstreams"
        ]
      }
    },
    "/api/v1/upstreams/{upstream_id}": {
      "delete": {
        "operationId": "delete_upstream_api_v1_upstreams__upstream_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "upstream_id",
            "required": true,
            "schema": {
              "title": "Upstream Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete an upstream",
        "tags": [
          "upstreams"
        ]
      },
      "patch": {
        "operationId": "update_upstream_api_v1_upstreams__upstream_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "upstream_id",
            "required": true,
            "schema": {
              "title": "Upstream Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpstreamUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpstreamRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update an upstream",
        "tags": [
          "upstreams"
        ]
      }
    },
...
  },
...
}